Added new 3D and PBC segmentation
Added a few new functions and many lines of code to enable a new 'box' 3D seeding approach and proper watershedding/adjustment of segmentation masks across periodic boundaries. The if-else split for the older 'column' seeding approach and new 'box' seeding approach should work (though I have not tested it), but right now the PBC treatment assumes we are using the 'box' approach and will likely have to be adapted to work with the 'column' approach too. There is also a bug in the buddy box seeding currently, as we assume a 5x5x5 seeding box the code will break if the size of the buddy box is less than 5 in any of the dimensions. Generalizing the box approach to be a user-set size and also adding a check for this in the Buddy Box routine should help to fix this, but I wanted to port it over without breaking any of the current functionality before I started exploring that.