Fixed wall_label overwrite procedure
Corrected an issue with multiple overwrites for the PBC wall_labels loop. Added new list "skip_list_thisind" to accumulate labels that have been dealt with during the course of treating the wall_label we are presently on. This ensures that we are not prematurely overwriting our present wall_label and exiting the loop through points for that label.