Added 3D segmentation with PBCs.  (#21)
* Added new 3D and PBC segmentation

Added a few new functions and many lines of code to enable a new 'box' 3D seeding approach and proper watershedding/adjustment of segmentation masks across periodic boundaries. The if-else split for the older 'column' seeding approach and new 'box' seeding approach should work (though I have not tested it), but right now the PBC treatment assumes we are using the 'box' approach and will likely have to be adapted to work with the 'column' approach too. There is also a bug in the buddy box seeding currently, as we assume a 5x5x5 seeding box the code will break if the size of the buddy box is less than 5 in any of the dimensions. Generalizing the box approach to be a user-set size and also adding a check for this in the Buddy Box routine should help to fix this, but I wanted to port it over without breaking any of the current functionality before I started exploring that.

* Moved some common functions to the utils module

* Removed deprecaded code, updated args

Removed older versions of segmentation functions which were nominally PBC/3D but had been superseded by new versions of the original functions including these capabilities and more. Also added 'seed_3D_flag' argument to all relevant functions for user choice of 3D feature seeding method

* Added missing seed_3D_flag

Added missing seed_3D_flag in segmentation function call for segmentation_3D function

* Starting to clean up

* continuing to clean up code, started process of adding tests

* More cleanup of segmentation, adding more tests to segmentation and feature detection

* Cleaning up of segmentation code, updating tests, cleaning up other code

* More cleaning up of segmentation

* Added compatibility of new PBC segmentation with 2D data

* fixed more bugs in segmentation with 2D inputs

* Adding more tests to PBC segmentation

* Added corner test case where it fails

* Potential fix to corner point bug

* Moved get_pbc_coordinates to utilities so that we can use it for segmentation ultimately

Also moved get_label_props_in_dict to inside the PBC flags as it's not used otherwise

* Updated docs, added start to test function for 3D box

* Added notes to myself

* Added comments and todos

* Update segmentation.py

added blah comment

* Revert "Update segmentation.py"

This reverts commit 22e22c21f831b3e9d194db9ce703e7db7a046d76.

* Added new comments explaining PBC seg approach

Added numerous comments pertaining to chunks of the PBC treatment procedure. Also found some likely code redundancies or deprecated pieces, which are noted in these new comments

* Updated non PBC seed box code

Updated the seed box code using my coordinates getter

* Added code coverage files to gitignore

* Segmentation now transposes to improve code readability

Refactored segmentation to transpose data internally so that z is always first (and then re-transposes after we finish). Added new tests to make sure that this is happening correctly

* Added new test to theoretically check buddy box

Note that none of my cases trigger the buddy box, which I'm honestly confused by.

* Got buddy boxes working for more cases

* Added new test that tests buddy box better.

Test currently failing, need to fix.

* Fixed wall_label overwrite procedure

Corrected an issue with multiple overwrites for the PBC wall_labels loop. Added new list "skip_list_thisind" to accumulate labels that have been dealt with during the course of treating the wall_label we are presently on. This ensures that we are not prematurely overwriting our present wall_label and exiting the loop through points for that label.

* Fixed buddy box segmentation bug.

* fixed github actions dockerfile location

Co-authored-by: galexsky <90701223+galexsky@users.noreply.github.com>
Co-authored-by: galexsky <galexsky94@gmail.com>