Bugfix on vertical coordinate implementation
I uncovered an issue for Iris Cubes with greater than 1 vertical coordinate (e.g. both 'altitude' and 'z' are provided). If this is the case and 'auto' is passed, utils.find_vertical_axis_from_coord asks you to specify the coordinate, however, if anything other than 1 or 'auto' was passed, the NotImplementedError in feature detection would be triggered. I also noticed that the vertical_coord and vertical_axis coordinates were not passed down the chain for the various layers of feature detection calls, and have changed this.