Tweak to vertical coord function and ncells fix
Slightly generalized and clarified value errors in find_vertical_axis_from_coord in utils module. Now, if default auto coordinate and more than one suitable coord is in iris cube, picks first of list. If cube lacks suitable coord for auto, the warning now specifies this. Additionally, if a coordinate is passed which for whatever reason does not exist in the cube, the warning also says this. Finally, a fix to ncells in segmentation which should assign the proper value at each index is implemented.